Update of sticky_finger.cpp to work with Gazebo 11
Only the one source files has been updated, the other two are not currently being used.  They do have elements that would require updating to compile.

The updates are enclosed with #define statements to allow the code to still compile against versions of Gazebo prior to 11.

Functional testing has not been conducted, but the updates are consistent with the Migrating.md document in Gazebo.